Make the contacts examples slightly more standalone
Instead of symlinking the scripts directory in the grantlee sources to
the source directory of this example, directly invoke the
extract_strings_linguist.py script with its relative path for the
example source directory.

The linguist_ts target still depends on something outside of it, however
now it is possible to build the contacts example on its own, even when
it is copied as standalone out of the grantlee sources.